feat: reset UI button
