+++ client/src/lib/const.ts
@@ -15,6 +15,7 @@ export const TILE_SIZE = 256;
 export const MIN_SCALE_FOR_DETAIL = 0.35;
 export const MIN_SCALE_FOR_ANIMATION = 0.35;
 export const DEFAULT_TIMEOUT = 30000; // Default timeout for waiting operations
+export const WIDGETS_STORAGE_KEY = 'ponziland-widgets-state';
 
 // Environment constants
 export const NAME_SPACE = 'ponzi_land';
